Some fixes to make the tool compile on additional compilers. Fixed a number of issues: - Evidently, the json library include files are [supposed](https://chromium.googlesource.com/external/github.com/open-source-parsers/jsoncpp/+/refs/heads/0.y.z/README.md) to be in the <json/*> directory not <jsoncpp/json/*> director. Ubuntu's location is added using a -I statement in the Makefile. - Removed a bunch of variables that are never used. - Initialized iid variable. (It was not initialized previously). - testResults.size() is of type unsigned long int. Cast the value to int.

+++ cpp/iid/iid_test_case.h
@@ -3,7 +3,7 @@
 
 #include <cstdlib>
 #include <vector> 
-#include <jsoncpp/json/json.h>
+#include <json/json.h>
 
 #include "permutation_test_result.h"
 #include "../shared/test_case_base.h"
@@ -47,7 +47,7 @@ public:
             json["h_i"] = h_i;
 
         Json::Value permutationTestResults;
-        for (int i = 0; i < testResults.size(); i++) {
+        for (int i = 0; i < (int)testResults.size(); i++) {
             permutationTestResults[i] = testResults[i].GetAsJson();
         }
